refactor: move SubMenuTriggerButton into own file
