+++ src/BrandedNavBar/SubMenuTrigger.tsx
@@ -1,12 +1,7 @@
 import React from "react";
-import styled, {
-  CSSObject,
-  StyledComponentPropsWithRef,
-} from "styled-components";
-import { Icon } from "../Icon";
-import { DropdownButton } from "../DropdownMenu/index";
 import NavBarDropdownMenu from "./NavBarDropdownMenu";
 import renderSubMenuItems from "./renderSubMenuItems";
+import SubMenuTriggerButton from "./SubMenuTriggerButton";
 
 type SubMenuTriggerProps = React.ComponentPropsWithRef<"button"> & {
   name?: string;
@@ -14,35 +9,6 @@ type SubMenuTriggerProps = React.ComponentPropsWithRef<"button"> & {
   onItemClick?: any;
   menuData: any[];
 };
-
-const StyledButton: StyledComponentPropsWithRef<any> = styled(DropdownButton)(
-  ({ isOpen, theme }: any): CSSObject => ({
-    position: "relative",
-    backgroundColor: isOpen ? theme.colors.lightBlue : "transparent",
-    color: isOpen ? theme.colors.darkBlue : theme.colors.darkGrey,
-  })
-);
-type SubMenuTriggerButtonProps = React.ComponentPropsWithRef<"button"> & {
-  name?: string;
-  isOpen: boolean;
-};
-
-const SubMenuTriggerButton = React.forwardRef<
-  HTMLButtonElement,
-  SubMenuTriggerButtonProps
->(({ name, isOpen, ...props }, ref) => (
-  <StyledButton isOpen={isOpen} ref={ref} {...props}>
-    {name}
-    <Icon
-      style={{ position: "absolute", top: "10px" }}
-      icon="rightArrow"
-      size="20px"
-      p="2px"
-    />
-  </StyledButton>
-));
-
-SubMenuTriggerButton.displayName = "SubMenuTriggerButton";
 
 const SubMenuTrigger = ({
   menuData,
